Refactor shortcake attributes to html attributes.
