+++ Example/Example/WordPressShortcode.swift
@@ -10,20 +10,9 @@ public struct Shortcode {
     }
 
     public let tag: String
-    public let attributes: ShortcodeAttributes
+    public let attributes: HTMLAttributes
     public let type: TagType
     public let content: String?
-}
-
-/// Struct to represent the attributes of a shortcode
-///
-public struct ShortcodeAttributes {
-
-    /// Attributes that have a form key=value or key="value" or key='value'
-    let namedAttributes: [String: String]
-
-    /// Attributes that have a form value "value"
-    let unamedAttributes: [String]
 }
 
 /// A class that processes a string and replace the designated shortcode for the replacement provided strings
@@ -69,9 +58,9 @@ open class ShortcodeProcessor: RegexProcessor {
             guard match.numberOfRanges == 8 else {
                 return nil
             }
-            var attributes = ShortcodeAttributes(namedAttributes: [:], unamedAttributes: [])
+            var attributes = HTMLAttributes(named: [:], unamed: [])
             if let attributesText = match.captureGroup(in:CaptureGroups.arguments.rawValue, text: text) {
-                attributes = ShortcodeAttributesParser.makeAttributes(in: attributesText)
+                attributes = HTMLAttributesParser.makeAttributes(in: attributesText)
             }
 
             var type: Shortcode.TagType = .single
@@ -89,74 +78,6 @@ open class ShortcodeProcessor: RegexProcessor {
 
         super.init(regex: regex, replacer: regexReplacer)
     }
-}
-
-
-/// A struct that parses attributes inside a shortcode and return the corresponding attributes object
-///
-public struct ShortcodeAttributesParser {
-
-    enum CaptureGroups: Int {
-        case all = 0
-        case nameInDoubleQuotes
-        case valueInDoubleQuotes
-        case nameInSingleQuotes
-        case valueInSingleQuotes
-        case nameUnquoted
-        case valueUnquoted
-        case justValueQuoted
-        case justValueUnquoted
-    }
-
-    /// Regular expression to detect attributes
-    /// This regular expression is reused from `shortcode_parse_atts()`
-    /// in `wp-includes/shortcodes.php`.
-    ///
-    /// Capture groups:
-    ///
-    /// 1. An attribute name, that corresponds to...
-    /// 2. a value in double quotes.
-    /// 3. An attribute name, that corresponds to...
-    /// 4. a value in single quotes.
-    /// 5. An attribute name, that corresponds to...
-    /// 6. an unquoted value.
-    /// 7. A numeric attribute in double quotes.
-    /// 8. An unquoted numeric attribute.
-    ///
-    static var attributesRegex: NSRegularExpression = {
-        let attributesPattern: String = "(\\w+)\\s*=\\s*\"([^\"]*)\"(?:\\s|$)|(\\w+)\\s*=\\s*'([^']*)'(?:\\s|$)|(\\w+)\\s*=\\s*([^\\s'\"]+)(?:\\s|$)|\"([^\"]*)\"(?:\\s|$)|(\\S+)(?:\\s|$)"
-        return try! NSRegularExpression(pattern: attributesPattern, options: .caseInsensitive)
-    }()
-
-    /// Parses the attributes from a string to the object
-    ///
-    /// - Parameter text: the text to where to find the attributes
-    /// - Returns: the ShortcodeAttributes parsed from the text
-    ///
-    static func makeAttributes(in text:String) -> ShortcodeAttributes {
-        var namedAttributes = [String: String]()
-        var unamedAttributes = [String]()
-
-        let attributesMatches = ShortcodeAttributesParser.attributesRegex.matches(in: text, options: [], range: text.nsRange(from: text.startIndex..<text.endIndex))
-        for attributeMatch in attributesMatches {
-            if let key = attributeMatch.captureGroup(in: CaptureGroups.nameInDoubleQuotes.rawValue, text: text),
-                let value = attributeMatch.captureGroup(in: CaptureGroups.valueInDoubleQuotes.rawValue, text: text) {
-                namedAttributes[key] = value
-            } else if let key = attributeMatch.captureGroup(in: CaptureGroups.nameInSingleQuotes.rawValue, text: text),
-                let value = attributeMatch.captureGroup(in: CaptureGroups.valueInSingleQuotes.rawValue, text: text) {
-                namedAttributes[key] = value
-            } else if let key = attributeMatch.captureGroup(in: CaptureGroups.nameUnquoted.rawValue, text: text),
-                let value = attributeMatch.captureGroup(in: CaptureGroups.valueUnquoted.rawValue, text: text) {
-                namedAttributes[key] = value
-            } else if let value = attributeMatch.captureGroup(in: CaptureGroups.justValueQuoted.rawValue, text: text) {
-                unamedAttributes.append(value)
-            } else if let value = attributeMatch.captureGroup(in: CaptureGroups.justValueUnquoted.rawValue, text: text) {
-                unamedAttributes.append(value)
-            }
-        }
-        return ShortcodeAttributes(namedAttributes: namedAttributes, unamedAttributes: unamedAttributes)
-    }
-
 }
 
 extension NSTextCheckingResult {
